Add files via upload
I developed  a project  Snake Game is a classic arcade game where the player controls a snake that moves around the screen, eating food and growing in size. The objective is to keep the snake alive by avoiding collisions with the walls or itself.This project is  made by using  Java, the Snake Game can be created using libraries like Swing for the graphical user interface (GUI) and AWT for handling key events and graphics rendering.